Add files via upload
This was the final assignment for my school project, I had to link the register/login page with SQL, then after registration, it had to pull info from SQL and send an email to the registered person with a unique code for verification before further login. I failed to do that part since I didn't understand it. But this is just a showcase of something I made 3 years ago.